requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
motor==3.3.2
pymongo==4.5.0
python-multipart==0.0.6
python-dotenv==1.0.0